+++ models/Kelas.php
@@ -123,6 +123,16 @@ class Kelas {
      */
     function delete($id){
         $id = $this->db->real_escape_string($id);
+
+        $cek = $this->db->query("
+            SELECT COUNT(*) as total FROM kelas_mahasiswa 
+            WHERE klsmhsKlsId = '$id' AND klsmhsIsAktif = 1
+        ")->fetch_assoc();
+
+        if ($cek['total'] > 0) {
+            throw new Exception("Kelas dengan Id $id tidak dapat dihapus karena masih digunakan difitur lain.");
+        }
+
         return $this->db->query("DELETE FROM kelas WHERE klsId = '$id'");
     }
